refactor: split i18n into a package
