+++ utils/i18n/__init__.py
@@ -1,0 +1,41 @@
+"""Localization helpers for lightweight UI string translation."""
+
+from __future__ import annotations
+
+from typing import Final, Literal
+
+from utils.i18n._en_us import MESSAGES as _EN_US
+from utils.i18n._pt_br import MESSAGES as _PT_BR
+
+LocaleCode = Literal["en-US", "pt-BR"]
+
+DEFAULT_LOCALE: Final[LocaleCode] = "en-US"
+SUPPORTED_LOCALES: Final[tuple[LocaleCode, ...]] = ("en-US", "pt-BR")
+
+LOCALE_LABELS: Final[dict[LocaleCode, str]] = {
+    "en-US": "English",
+    "pt-BR": "Português (Brasil)",
+}
+
+MESSAGES: Final[dict[LocaleCode, dict[str, str]]] = {
+    "en-US": _EN_US,
+    "pt-BR": _PT_BR,
+}
+
+
+def normalize_locale(locale: str | None) -> LocaleCode:
+    """Normalize locale values to the supported locale set."""
+    if locale in SUPPORTED_LOCALES:
+        return locale
+    return DEFAULT_LOCALE
+
+
+def translate(locale: str | None, key: str, **kwargs: object) -> str:
+    """Return a translated string with fallback to default locale and key text."""
+    normalized = normalize_locale(locale)
+    template = MESSAGES.get(normalized, {}).get(key) or MESSAGES[DEFAULT_LOCALE].get(key)
+    if template is None:
+        return key
+    if not kwargs:
+        return template
+    return template.format(**kwargs)
