+++ utils/i18n/_en_us.py
@@ -1,0 +1,287 @@
+"""English (United States) locale strings."""
+
+MESSAGES: dict[str, str] = {
+    "app.status.ready": "Ready",
+    "app.status.language_changed": "Language changed. Restart app to fully apply it.",
+    "app.status.selected_language": "Selected language: {language}",
+    "app.title.main_frame": "MTGO Deck Research & Builder",
+    "app.label.deck_data_source": "Deck data source:",
+    "app.label.language": "Language:",
+    "app.menu.deck_data_source": "Deck data source",
+    "app.menu.language": "Language",
+    "app.choice.source.both": "Both",
+    "app.choice.source.mtggoldfish": "MTGGoldfish",
+    "app.choice.source.mtgo": "MTGO.com",
+    "toolbar.opponent_tracker": "Opponent Tracker",
+    "toolbar.timer_alert": "Timer Alert",
+    "toolbar.match_history": "Match History",
+    "toolbar.metagame_analysis": "Metagame Analysis",
+    "toolbar.settings": "Settings",
+    "toolbar.load_collection": "Load Collection",
+    "toolbar.download_card_images": "Download Card Images",
+    "toolbar.update_card_database": "Update Card Database",
+    "toolbar.export_diagnostics": "Export Diagnostics",
+    "toolbar.show_tutorial": "Show Tutorial",
+    "toolbar.help": "Help (F1)",
+    "toolbar.tooltip.opponent_tracker": "Detect your current MTGO opponent and look up their most-played archetypes",
+    "toolbar.tooltip.timer_alert": "Set a countdown timer alert to warn you before round time runs out",
+    "toolbar.tooltip.match_history": "Parse your MTGO GameLog files and view recent match results",
+    "toolbar.tooltip.metagame_analysis": "Browse the format metagame breakdown and archetype share data",
+    "deck_actions.daily_average": "Today's Average",
+    "deck_actions.copy": "Copy",
+    "deck_actions.load_deck": "Load Deck",
+    "deck_actions.save_deck": "Save Deck",
+    "deck_actions.tooltip.daily_average": "Build an averaged deck from today's tournament results",
+    "deck_actions.tooltip.copy": "Copy the current deck list to your clipboard",
+    "deck_actions.tooltip.load_deck": "Load a saved deck from file or database",
+    "deck_actions.tooltip.save_deck": "Save the current deck",
+    "research.format": "Format",
+    "research.search_hint": "Search archetypes...",
+    "research.reload_archetypes": "Reload Archetypes",
+    "research.loading_archetypes": "Loading...",
+    "research.failed_archetypes": "Failed to load archetypes.",
+    "research.no_archetypes": "No archetypes found.",
+    "research.switch_to_builder": "Deck Builder",
+    "research.tooltip.format": "Select the format to research",
+    "research.tooltip.search": "Filter the archetype list by name",
+    "research.tooltip.archetypes": "Click an archetype to load its decklists",
+    "research.tooltip.reload": "Refresh archetype data from MTGGoldfish",
+    "tabs.mainboard": "Mainboard",
+    "tabs.sideboard": "Sideboard",
+    "tabs.sideboard_guide": "Sideboard Guide",
+    "tabs.deck_notes": "Deck Notes",
+    "tabs.tooltip.mainboard": "Your main deck (typically 60 cards)",
+    "tabs.tooltip.sideboard": "Your sideboard (up to 15 cards)",
+    "tabs.tooltip.sideboard_guide": "Matchup-by-matchup sideboard plan and notes",
+    "tabs.tooltip.deck_notes": "Free-form notes attached to this deck",
+    "app.label.deck_workspace": "Deck Workspace",
+    "app.label.deck_results": "Deck Results",
+    "app.label.card_inspector": "Card Inspector",
+    "app.label.left_panel.research": "Research",
+    "app.label.left_panel.builder": "Builder",
+    "app.status.collection_not_loaded": "Collection inventory not loaded.",
+    "app.status.select_archetype": "Select an archetype to view decks.",
+    "deck_results.status.loaded_decks": "Loaded {count} decks for {archetype}. Click a deck to load it.",
+    "deck_results.total_loaded": "Total decks loaded: {count}",
+    "deck_results.recent_activity": "Recent activity:",
+    "deck_results.no_activity": "No recent deck activity.",
+    "deck_results.no_decks": "No decks found.",
+    "deck_results.no_decks_for": "No decks for {archetype}.",
+    "deck_results.failed_load": "Failed to load decks.",
+    "app.status.card_db_loading": "Loading card database...",
+    "app.status.card_db_loaded": "Card database loaded",
+    "app.status.card_db_failed": "Card database load failed",
+    "app.collection.not_found": "No collection found. Click \u2018Refresh Collection\u2019 to fetch from MTGO.",
+    "app.research.archetypes_loaded": "Loaded {count} archetypes for {format}.",
+    "app.research.select_archetype_loaded": "Select an archetype to view decks.\nLoaded {count} archetypes.",
+    "builder.back_button": "Deck Research",
+    "builder.back_button.tooltip": "Switch back to Deck Research mode",
+    "builder.info": "Deck Builder: search MTG cards by property.",
+    "builder.field.card_name": "Card Name",
+    "builder.field.type_line": "Type Line",
+    "builder.field.mana_cost": "Mana Cost",
+    "builder.field.oracle_text": "Oracle Text",
+    "builder.field.mana_value": "Mana Value Filter",
+    "builder.filter.color_identity": "Color Identity Filter",
+    "builder.filter.format": "Format",
+    "builder.clear_filters": "Clear Filters",
+    "builder.radar.use_filter": "Use Radar Filter",
+    "builder.radar.open": "Open Radar...",
+    "builder.add_to_main": "+ Mainboard",
+    "builder.add_to_side": "+ Sideboard",
+    "builder.status.results": "Results update automatically as you type.",
+    "builder.col.name": "Name",
+    "builder.col.mana_cost": "Mana Cost",
+    "builder.hint.card_name": "e.g. Ragavan",
+    "builder.hint.type_line": "Artifact Creature",
+    "builder.hint.mana_cost": "Curly braces like {1}{G} or shorthand (e.g. GGG)",
+    "builder.label.match": "Match",
+    "builder.check.exact_symbols": "Exact symbols",
+    "builder.btn.adv_filters_show": "+ Advanced Filters",
+    "builder.btn.adv_filters_hide": "- Advanced Filters",
+    "builder.hint.oracle_text": "Keywords or abilities",
+    "builder.hint.mana_value": "e.g. 3",
+    "builder.format.any": "Any",
+    "guide.col.archetype": "Archetype",
+    "guide.col.play_out": "Play: Out",
+    "guide.col.play_in": "Play: In",
+    "guide.col.draw_out": "Draw: Out",
+    "guide.col.draw_in": "Draw: In",
+    "guide.col.notes": "Notes",
+    "guide.btn.add": "Add",
+    "guide.btn.edit": "Edit",
+    "guide.btn.delete": "Delete",
+    "guide.btn.exclusions": "Exclusions",
+    "guide.btn.export": "Export",
+    "guide.btn.import": "Import",
+    "guide.btn.flex_slots": "Flex Slots",
+    "guide.btn.pin": "Pin for Tracker",
+    "guide.btn.pinned": "Pinned \u2713",
+    "guide.btn.cta": "Add your first matchup",
+    "guide.empty": "No matchup notes yet.\nClick \u201cAdd\u201d to create a sideboard guide entry.",
+    "guide.label.exclusions": "Exclusions",
+    "guide.tooltip.flex_slots": (
+        "Mark mainboard cards as flex slots. Flex slots are highlighted in the Out selectors "
+        "when creating guide entries, making it easier to identify cards to side out."
+    ),
+    "guide.dialog.archetype_matchup": "Archetype/Matchup",
+    "guide.dialog.cancel": "Cancel",
+    "guide.dialog.on_the_play": "ON THE PLAY",
+    "guide.dialog.on_the_draw": "ON THE DRAW",
+    "guide.dialog.out_from_main": "Out (from Mainboard)",
+    "guide.dialog.in_from_side": "In (from Sideboard)",
+    "guide.dialog.notes_label": "Notes (Optional)",
+    "guide.dialog.notes_hint": "Strategy notes for this matchup",
+    "guide.dialog.double_entries": "Enable double entries",
+    "guide.dialog.save_continue": "Save & Continue",
+    "guide.selector.cards_selected": "{count} cards selected",
+    "radar.label.no_radar": "No radar loaded.",
+    "radar.btn.export": "Export as Decklist",
+    "radar.btn.use_search": "Use for Search",
+    "radar.box.mainboard": "Mainboard Radar",
+    "radar.box.sideboard": "Sideboard Radar",
+    "radar.col.card": "Card",
+    "radar.col.inclusion": "Inclusion %",
+    "radar.col.expected": "Expected Copies",
+    "radar.col.avg": "Avg Copies",
+    "radar.col.max": "Max",
+    "radar.dialog.select_archetype": "Select Archetype:",
+    "radar.dialog.generate": "Generate Radar",
+    "radar.btn.cancel": "Cancel",
+    "radar.btn.close": "Close",
+    "notes.btn.add": "+ Add Note",
+    "notes.btn.save": "Save Notes",
+    "notes.empty": "No deck notes yet, click \u201cAdd\u201d to create a deck note entry.",
+    "notes.saved": "Deck notes saved.",
+    "notes.type.general": "General",
+    "notes.type.matchup": "Matchup",
+    "notes.type.sideboard_plan": "Sideboard Plan",
+    "notes.type.custom": "Custom",
+    "tracker.label.not_detected": "Opponent not detected",
+    "tracker.label.watching": "Watching for MTGO match windows\u2026",
+    "tracker.btn.refresh": "Refresh",
+    "tracker.btn.calculator": "Calculator",
+    "tracker.btn.guide": "Guide",
+    "tracker.btn.close": "Close",
+    "tracker.status.no_active_match": "No active match detected",
+    "tracker.status.waiting": "Waiting for MTGO match window\u2026",
+    "metagame.chart.no_data": "No data available for selected period",
+    "metagame.label.format": "Format:",
+    "metagame.label.time_window": "Time Window (days):",
+    "metagame.label.starting_from": "Starting from day:",
+    "metagame.label.changes": "Metagame Changes",
+    "metagame.btn.refresh": "Refresh Data",
+    "metagame.loaded": "Loaded {count} archetypes",
+    "metagame.period.last_days": "Last {count} day(s)",
+    "metagame.period.days_ago": "{count} day(s) ago",
+    "metagame.period.range_days_ago": "{start}-{end} days ago",
+    "metagame.changes.no_data": "No comparison data available",
+    "metagame.changes.vs_period": "Changes vs {period}",
+    "metagame.changes.none": "No significant changes",
+    "metagame.status.fetching": "Fetching metagame data...",
+    "metagame.status.error": "Unable to load metagame data:\n{message}",
+    "match.metrics.title": "Win-Rate Metrics",
+    "match.metrics.abs_match_rate": "Absolute Match Win Rate",
+    "match.metrics.abs_game_rate": "Absolute Game Win Rate",
+    "match.metrics.filtered_match_rate": "Match Win Rate (filtered)",
+    "match.metrics.filtered_game_rate": "Game Win Rate (filtered)",
+    "match.metrics.mulligan_rate": "Mulligan Rate",
+    "match.metrics.avg_mulligans": "Avg Mulligans/Match",
+    "match.metrics.opp_match_rate": "Vs. Opponent Match Win Rate",
+    "match.metrics.opp_mull_rate": "Vs. Opponent Mull Rate",
+    "match.filter.start": "Start (YYYY-MM-DD):",
+    "match.filter.end": "End (YYYY-MM-DD):",
+    "match.filter.apply": "Apply Date Filter",
+    "match.col.players": "Players (Archetypes)",
+    "match.col.result": "Result",
+    "match.col.mulligans": "Mulligans",
+    "match.col.date": "Date",
+    "match.btn.refresh": "Refresh",
+    "match.status.loading": "Loading all match history...",
+    "match.status.parsing": "Parsing {current}/{total} matches...",
+    "match.status.loaded": "Loaded {count} matches",
+    "match.status.failed": "Failed to load match history.",
+    "match.status.no_data": "No match data available.",
+    "match.status.invalid_date": "Invalid date format",
+    "match.result.won": "Won",
+    "match.result.lost": "Lost",
+    "timer.section.thresholds": "Alert Thresholds",
+    "timer.section.challenge": "Active Challenge Timer",
+    "timer.label.sound": "Alert Sound:",
+    "timer.label.check_interval": "Check interval (ms):",
+    "timer.label.repeat_interval": "Repeat interval (seconds):",
+    "timer.check.start_alert": "Alert when timer starts counting down",
+    "timer.check.repeat_alarm": "Repeat alarm at interval",
+    "timer.btn.start": "Start Monitoring",
+    "timer.btn.stop": "Stop",
+    "timer.btn.test": "Test Alert",
+    "timer.no_challenge": "No active challenge timer detected.",
+    "timer.configure": "Configure thresholds and click Start to begin monitoring.",
+    "tutorial.dialog_title": "MTGO Tools \u2014 Quick Tour",
+    "tutorial.btn.skip": "Skip Tour",
+    "tutorial.btn.back": "< Back",
+    "tutorial.btn.next": "Next >",
+    "tutorial.btn.finish": "Finish",
+    "tutorial.step0.title": "Welcome to MTGO Tools",
+    "tutorial.step0.body": (
+        "MTGO Tools helps you research the competitive metagame, build and edit decks, "
+        "track opponents, and manage your MTGO collection \u2014 all in one desktop app.\n\n"
+        "This short tour covers the main features. You can revisit it any time from "
+        "Settings \u2192 Show Tutorial."
+    ),
+    "tutorial.step1.title": "Metagame Research",
+    "tutorial.step1.body": (
+        "The left panel is your metagame research hub.\n\n"
+        "\u2022  Choose a format (Modern, Legacy, \u2026) from the dropdown.\n"
+        "\u2022  Type in the search box to filter archetypes by name.\n"
+        "\u2022  Click an archetype to load its decklists in the Deck Results panel.\n"
+        "\u2022  Use \u201cReload Archetypes\u201d to refresh data from MTGGoldfish."
+    ),
+    "tutorial.step2.title": "Deck Workspace",
+    "tutorial.step2.body": (
+        "The centre area shows the currently loaded deck.\n\n"
+        "\u2022  Mainboard \u2014 your 60-card main deck.\n"
+        "\u2022  Sideboard \u2014 your 15-card sideboard.\n"
+        "\u2022  Hover over or click a card row to inspect it in the Card Inspector "
+        "on the right.\n"
+        "\u2022  Use the + / \u2212 controls to edit counts when building your own deck."
+    ),
+    "tutorial.step3.title": "Toolbar Tools",
+    "tutorial.step3.body": (
+        "The toolbar at the top of the right panel provides quick access to:\n\n"
+        "\u2022  Opponent Tracker \u2014 detects the opponent from your MTGO window title "
+        "and looks up their most-played archetypes.\n"
+        "\u2022  Timer Alert \u2014 configurable countdown to warn you before time runs out "
+        "in a round.\n"
+        "\u2022  Match History \u2014 parses your MTGO GameLog files and shows recent results.\n"
+        "\u2022  Metagame Analysis \u2014 a top-level breakdown of the current format."
+    ),
+    "tutorial.step4.title": "Deck Builder",
+    "tutorial.step4.body": (
+        "Switch the left panel to Builder mode to search for cards and craft your own deck.\n\n"
+        "\u2022  Type a card name or keyword in the search box.\n"
+        "\u2022  Click a result to preview it in the Card Inspector.\n"
+        "\u2022  Use \u201cAdd to Main\u201d or \u201cAdd to Side\u201d to add it to your deck.\n"
+        "\u2022  Open the Mana Keyboard for quick mana-cost symbol input.\n"
+        "\u2022  Use \u201cCopy\u201d to copy the deck list to your clipboard."
+    ),
+    "tutorial.step5.title": "Sideboard Guide",
+    "tutorial.step5.body": (
+        "The Sideboard Guide tab lets you record matchup-by-matchup notes.\n\n"
+        "\u2022  Add an entry for each archetype you face.\n"
+        "\u2022  Record cards to bring IN and take OUT for each matchup.\n"
+        "\u2022  Mark flex slots \u2014 cards whose count varies by matchup.\n"
+        "\u2022  Pin the guide to keep it visible while reviewing other tabs.\n"
+        "\u2022  Export or import as CSV to share guides with teammates."
+    ),
+    "tutorial.step6.title": "You\u2019re All Set!",
+    "tutorial.step6.body": (
+        "That\u2019s the quick tour of MTGO Tools.\n\n"
+        "A few more tips:\n"
+        "\u2022  Use the \u2699 Settings menu to load your MTGO collection, download card "
+        "images, update the card database, or change the language.\n"
+        "\u2022  Deck Notes let you keep free-form notes attached to any deck.\n"
+        "\u2022  Session state (current deck, format, window size) is saved automatically.\n\n"
+        "Good luck in your matches!"
+    ),
+}
