+++ utils/i18n/_pt_br.py
@@ -1,0 +1,285 @@
+"""Portuguese (Brazil) locale strings."""
+
+MESSAGES: dict[str, str] = {
+    "app.status.ready": "Pronto",
+    "app.status.language_changed": "Idioma alterado. Reinicie o app para aplicar por completo.",
+    "app.status.selected_language": "Idioma selecionado: {language}",
+    "app.title.main_frame": "Pesquisa e Montagem de Deck MTGO",
+    "app.label.deck_data_source": "Fonte de decks:",
+    "app.label.language": "Idioma:",
+    "app.menu.deck_data_source": "Fonte de decks",
+    "app.menu.language": "Idioma",
+    "app.choice.source.both": "Ambos",
+    "app.choice.source.mtggoldfish": "MTGGoldfish",
+    "app.choice.source.mtgo": "MTGO.com",
+    "toolbar.opponent_tracker": "Rastreador de Oponente",
+    "toolbar.timer_alert": "Alerta de Tempo",
+    "toolbar.match_history": "Histórico de Partidas",
+    "toolbar.metagame_analysis": "Análise de Metagame",
+    "toolbar.settings": "Configurações",
+    "toolbar.load_collection": "Carregar Coleção",
+    "toolbar.download_card_images": "Baixar Imagens de Cartas",
+    "toolbar.update_card_database": "Atualizar Banco de Cartas",
+    "toolbar.export_diagnostics": "Exportar Diagnóstico",
+    "toolbar.show_tutorial": "Mostrar Tutorial",
+    "toolbar.help": "Ajuda (F1)",
+    "toolbar.tooltip.opponent_tracker": "Detecta seu oponente atual no MTGO e busca os arquétipos mais jogados por ele",
+    "toolbar.tooltip.timer_alert": "Define um alerta de contagem regressiva antes do tempo da rodada acabar",
+    "toolbar.tooltip.match_history": "Analisa seus arquivos GameLog do MTGO e exibe resultados recentes",
+    "toolbar.tooltip.metagame_analysis": "Veja a distribuição do metagame e a participação de cada arquétipo no formato",
+    "deck_actions.daily_average": "Média de Hoje",
+    "deck_actions.copy": "Copiar",
+    "deck_actions.load_deck": "Carregar Deck",
+    "deck_actions.save_deck": "Salvar Deck",
+    "deck_actions.tooltip.daily_average": "Monta um deck com a média dos torneios de hoje",
+    "deck_actions.tooltip.copy": "Copia a lista do deck atual para a área de transferência",
+    "deck_actions.tooltip.load_deck": "Carrega um deck salvo de arquivo ou banco de dados",
+    "deck_actions.tooltip.save_deck": "Salva o deck atual",
+    "research.format": "Formato",
+    "research.search_hint": "Buscar arquétipos...",
+    "research.reload_archetypes": "Recarregar Arquétipos",
+    "research.loading_archetypes": "Carregando...",
+    "research.failed_archetypes": "Falha ao carregar arquétipos.",
+    "research.no_archetypes": "Nenhum arquétipo encontrado.",
+    "research.switch_to_builder": "Montador de Deck",
+    "research.tooltip.format": "Selecione o formato para pesquisar",
+    "research.tooltip.search": "Filtre a lista de arquétipos por nome",
+    "research.tooltip.archetypes": "Clique em um arquétipo para carregar seus decklists",
+    "research.tooltip.reload": "Atualiza os dados de arquétipos do MTGGoldfish",
+    "tabs.mainboard": "Principal",
+    "tabs.sideboard": "Sideboard",
+    "tabs.sideboard_guide": "Guia de Sideboard",
+    "tabs.deck_notes": "Notas de Deck",
+    "tabs.tooltip.mainboard": "Seu deck principal (geralmente 60 cartas)",
+    "tabs.tooltip.sideboard": "Seu sideboard (até 15 cartas)",
+    "tabs.tooltip.sideboard_guide": "Plano de sideboard e notas por matchup",
+    "tabs.tooltip.deck_notes": "Notas em texto livre para este deck",
+    "app.label.deck_workspace": "Área de Trabalho",
+    "app.label.deck_results": "Resultados de Deck",
+    "app.label.card_inspector": "Inspetor de Cartas",
+    "app.label.left_panel.research": "Pesquisa",
+    "app.label.left_panel.builder": "Montador",
+    "app.status.collection_not_loaded": "Inventário de coleção não carregado.",
+    "app.status.select_archetype": "Selecione um arquétipo para ver os decks.",
+    "deck_results.status.loaded_decks": "Carregados {count} decks para {archetype}. Clique em um deck para carregá-lo.",
+    "deck_results.total_loaded": "Total de decks carregados: {count}",
+    "deck_results.recent_activity": "Atividade recente:",
+    "deck_results.no_activity": "Nenhuma atividade recente de decks.",
+    "deck_results.no_decks": "Nenhum deck encontrado.",
+    "deck_results.no_decks_for": "Nenhum deck para {archetype}.",
+    "deck_results.failed_load": "Falha ao carregar decks.",
+    "app.status.card_db_loading": "Carregando banco de dados de cartas...",
+    "app.status.card_db_loaded": "Banco de dados de cartas carregado",
+    "app.status.card_db_failed": "Falha ao carregar banco de dados de cartas",
+    "app.collection.not_found": "Nenhuma coleção encontrada. Clique em \u2018Atualizar Coleção\u2019 para buscar do MTGO.",
+    "app.research.archetypes_loaded": "Carregados {count} arquétipos para {format}.",
+    "app.research.select_archetype_loaded": "Selecione um arquétipo para ver os decks.\nCarregados {count} arquétipos.",
+    "builder.back_button": "Pesquisa de Deck",
+    "builder.back_button.tooltip": "Voltar ao modo de Pesquisa de Deck",
+    "builder.info": "Montador de Deck",
+    "builder.field.card_name": "Nome da Carta",
+    "builder.field.type_line": "Tipo",
+    "builder.field.mana_cost": "Custo de Mana",
+    "builder.field.oracle_text": "Texto",
+    "builder.field.mana_value": "Valor de Mana",
+    "builder.filter.color_identity": "Identidade de Cor",
+    "builder.filter.format": "Formato",
+    "builder.clear_filters": "Limpar Filtros",
+    "builder.radar.use_filter": "Usar Filtro de Radar",
+    "builder.radar.open": "Abrir Radar...",
+    "builder.add_to_main": "+ Principal",
+    "builder.add_to_side": "+ Sideboard",
+    "builder.status.results": "Os resultados atualizam automaticamente enquanto você digita.",
+    "builder.col.name": "Nome",
+    "builder.col.mana_cost": "Custo de Mana",
+    "builder.hint.card_name": "ex.: Ragavan",
+    "builder.hint.type_line": "Criatura Artefato",
+    "builder.hint.mana_cost": "Chaves como {1}{V} ou atalho (ex.: VVV)",
+    "builder.label.match": "Combinar",
+    "builder.check.exact_symbols": "Símbolos exatos",
+    "builder.btn.adv_filters_show": "+ Filtros Avançados",
+    "builder.btn.adv_filters_hide": "- Filtros Avançados",
+    "builder.hint.oracle_text": "Palavras-chave ou habilidades",
+    "builder.hint.mana_value": "ex.: 3",
+    "builder.format.any": "Qualquer",
+    "guide.col.archetype": "Arquétipo",
+    "guide.col.play_out": "Jogar: Sai",
+    "guide.col.play_in": "Jogar: Entra",
+    "guide.col.draw_out": "Comprar: Sai",
+    "guide.col.draw_in": "Comprar: Entra",
+    "guide.col.notes": "Notas",
+    "guide.btn.add": "Adicionar",
+    "guide.btn.edit": "Editar",
+    "guide.btn.delete": "Excluir",
+    "guide.btn.exclusions": "Exclusões",
+    "guide.btn.export": "Exportar",
+    "guide.btn.import": "Importar",
+    "guide.btn.flex_slots": "Flex Slots",
+    "guide.btn.pin": "Rastrear",
+    "guide.btn.pinned": "Fixado \u2713",
+    "guide.btn.cta": "Adicionar primeiro matchup",
+    "guide.empty": "Nenhuma nota de matchup ainda.\nClique em \u201cAdicionar\u201d para criar uma entrada.",
+    "guide.label.exclusions": "Exclusões",
+    "guide.tooltip.flex_slots": (
+        "Marque cartas do mainboard como flex slots. Flex slots são destacados nos seletores de saída "
+        "ao criar entradas de guia, facilitando a identificação das cartas a serem tiradas."
+    ),
+    "guide.dialog.archetype_matchup": "Arquétipo/Matchup",
+    "guide.dialog.cancel": "Cancelar",
+    "guide.dialog.on_the_play": "NO PRIMEIRO TURNO",
+    "guide.dialog.on_the_draw": "NO SEGUNDO TURNO",
+    "guide.dialog.out_from_main": "Saindo (do Mainboard)",
+    "guide.dialog.in_from_side": "Entrando (do Sideboard)",
+    "guide.dialog.notes_label": "Notas (Opcional)",
+    "guide.dialog.notes_hint": "Notas de estratégia para este matchup",
+    "guide.dialog.double_entries": "Habilitar entradas duplas",
+    "guide.dialog.save_continue": "Salvar e Continuar",
+    "guide.selector.cards_selected": "{count} cartas selecionadas",
+    "radar.label.no_radar": "Nenhum radar carregado.",
+    "radar.btn.export": "Exportar como Decklist",
+    "radar.btn.use_search": "Usar na Pesquisa",
+    "radar.box.mainboard": "Radar do Mainboard",
+    "radar.box.sideboard": "Radar do Sideboard",
+    "radar.col.card": "Carta",
+    "radar.col.inclusion": "% Inclusão",
+    "radar.col.expected": "Cópias Esperadas",
+    "radar.col.avg": "Cópias Médias",
+    "radar.col.max": "Máx",
+    "radar.dialog.select_archetype": "Selecionar Arquétipo:",
+    "radar.dialog.generate": "Gerar Radar",
+    "radar.btn.cancel": "Cancelar",
+    "radar.btn.close": "Fechar",
+    "notes.btn.add": "+ Adicionar Nota",
+    "notes.btn.save": "Salvar Notas",
+    "notes.empty": "Nenhuma nota de deck ainda, clique em \u201cAdicionar\u201d para criar uma nota.",
+    "notes.saved": "Notas de deck salvas.",
+    "notes.type.general": "Geral",
+    "notes.type.matchup": "Matchup",
+    "notes.type.sideboard_plan": "Plano de Sideboard",
+    "notes.type.custom": "Personalizado",
+    "tracker.label.not_detected": "Oponente não detectado",
+    "tracker.label.watching": "Monitorando janelas de partida MTGO\u2026",
+    "tracker.btn.refresh": "Atualizar",
+    "tracker.btn.calculator": "Calculadora",
+    "tracker.btn.guide": "Guia",
+    "tracker.btn.close": "Fechar",
+    "tracker.status.no_active_match": "Nenhuma partida ativa detectada",
+    "tracker.status.waiting": "Aguardando janela de partida MTGO\u2026",
+    "metagame.chart.no_data": "Sem dados disponíveis para o período selecionado",
+    "metagame.label.format": "Formato:",
+    "metagame.label.time_window": "Janela de Tempo (dias):",
+    "metagame.label.starting_from": "A partir do dia:",
+    "metagame.label.changes": "Mudanças de Metagame",
+    "metagame.btn.refresh": "Atualizar Dados",
+    "metagame.loaded": "Carregados {count} arquétipos",
+    "metagame.period.last_days": "Últimos {count} dia(s)",
+    "metagame.period.days_ago": "há {count} dia(s)",
+    "metagame.period.range_days_ago": "há {start}-{end} dias",
+    "metagame.changes.no_data": "Sem dados de comparação disponíveis",
+    "metagame.changes.vs_period": "Mudanças em relação a {period}",
+    "metagame.changes.none": "Nenhuma mudança significativa",
+    "metagame.status.fetching": "Buscando dados de metagame...",
+    "metagame.status.error": "Não foi possível carregar dados de metagame:\n{message}",
+    "match.metrics.title": "Métricas de Taxa de Vitória",
+    "match.metrics.abs_match_rate": "Taxa Absoluta de Vitórias em Partidas",
+    "match.metrics.abs_game_rate": "Taxa Absoluta de Vitórias em Jogos",
+    "match.metrics.filtered_match_rate": "Taxa de Vitórias em Partidas (filtrado)",
+    "match.metrics.filtered_game_rate": "Taxa de Vitórias em Jogos (filtrado)",
+    "match.metrics.mulligan_rate": "Taxa de Mulligans",
+    "match.metrics.avg_mulligans": "Média de Mulligans/Partida",
+    "match.metrics.opp_match_rate": "Taxa de Vitórias vs. Oponente",
+    "match.metrics.opp_mull_rate": "Taxa de Mulligans vs. Oponente",
+    "match.filter.start": "Início (AAAA-MM-DD):",
+    "match.filter.end": "Fim (AAAA-MM-DD):",
+    "match.filter.apply": "Aplicar Filtro de Data",
+    "match.col.players": "Jogadores (Arquétipos)",
+    "match.col.result": "Resultado",
+    "match.col.mulligans": "Mulligans",
+    "match.col.date": "Data",
+    "match.btn.refresh": "Atualizar",
+    "match.status.loading": "Carregando todo o histórico de partidas...",
+    "match.status.parsing": "Processando {current}/{total} partidas...",
+    "match.status.loaded": "Carregadas {count} partidas",
+    "match.status.failed": "Falha ao carregar histórico de partidas.",
+    "match.status.no_data": "Nenhum dado de partida disponível.",
+    "match.status.invalid_date": "Formato de data inválido",
+    "match.result.won": "Vitória",
+    "match.result.lost": "Derrota",
+    "timer.section.thresholds": "Limites de Alerta",
+    "timer.section.challenge": "Cronômetro de Desafio Ativo",
+    "timer.label.sound": "Som do Alerta:",
+    "timer.label.check_interval": "Intervalo de verificação (ms):",
+    "timer.label.repeat_interval": "Intervalo de repetição (segundos):",
+    "timer.check.start_alert": "Alertar quando o cronômetro iniciar contagem regressiva",
+    "timer.check.repeat_alarm": "Repetir alarme no intervalo",
+    "timer.btn.start": "Iniciar Monitoramento",
+    "timer.btn.stop": "Parar",
+    "timer.btn.test": "Testar Alerta",
+    "timer.no_challenge": "Nenhum cronômetro de desafio ativo detectado.",
+    "timer.configure": "Configure os limites e clique em Iniciar para começar a monitorar.",
+    "tutorial.dialog_title": "MTGO Tools \u2014 Tour R\u00e1pido",
+    "tutorial.btn.skip": "Pular Tour",
+    "tutorial.btn.back": "< Voltar",
+    "tutorial.btn.next": "Pr\u00f3ximo >",
+    "tutorial.btn.finish": "Concluir",
+    "tutorial.step0.title": "Bem-vindo ao MTGO Tools",
+    "tutorial.step0.body": (
+        "O MTGO Tools ajuda voc\u00ea a pesquisar o metagame competitivo, montar e editar decks, "
+        "rastrear oponentes e gerenciar sua cole\u00e7\u00e3o MTGO \u2014 tudo em um \u00fanico aplicativo de desktop.\n\n"
+        "Este pequeno tour apresenta os principais recursos. Voc\u00ea pode revisit\u00e1-lo a qualquer momento em "
+        "Configura\u00e7\u00f5es \u2192 Mostrar Tutorial."
+    ),
+    "tutorial.step1.title": "Pesquisa de Metagame",
+    "tutorial.step1.body": (
+        "O painel esquerdo \u00e9 seu centro de pesquisa de metagame.\n\n"
+        "\u2022  Escolha um formato (Modern, Legacy, \u2026) no menu suspenso.\n"
+        "\u2022  Digite na caixa de busca para filtrar arqu\u00e9tipos por nome.\n"
+        "\u2022  Clique em um arqu\u00e9tipo para carregar seus decklists no painel de Resultados.\n"
+        "\u2022  Use \u201cRecarregar Arqu\u00e9tipos\u201d para atualizar os dados do MTGGoldfish."
+    ),
+    "tutorial.step2.title": "\u00c1rea de Deck",
+    "tutorial.step2.body": (
+        "A \u00e1rea central mostra o deck carregado atualmente.\n\n"
+        "\u2022  Principal \u2014 seu deck principal de 60 cartas.\n"
+        "\u2022  Sideboard \u2014 seu sideboard de 15 cartas.\n"
+        "\u2022  Passe o mouse ou clique em uma linha de carta para inspecion\u00e1-la no Inspetor de Cartas \u00e0 direita.\n"
+        "\u2022  Use os controles + / \u2212 para editar quantidades ao montar seu pr\u00f3prio deck."
+    ),
+    "tutorial.step3.title": "Ferramentas da Barra",
+    "tutorial.step3.body": (
+        "A barra de ferramentas no topo do painel direito oferece acesso r\u00e1pido a:\n\n"
+        "\u2022  Rastreador de Oponente \u2014 detecta o oponente pelo t\u00edtulo da janela do MTGO "
+        "e busca os arqu\u00e9tipos mais jogados por ele.\n"
+        "\u2022  Alerta de Tempo \u2014 contagem regressiva configur\u00e1vel para avisar antes do tempo acabar na rodada.\n"
+        "\u2022  Hist\u00f3rico de Partidas \u2014 analisa seus arquivos GameLog do MTGO e mostra resultados recentes.\n"
+        "\u2022  An\u00e1lise de Metagame \u2014 vis\u00e3o geral do formato atual."
+    ),
+    "tutorial.step4.title": "Montador de Deck",
+    "tutorial.step4.body": (
+        "Alterne o painel esquerdo para o modo Montador para buscar cartas e criar seu pr\u00f3prio deck.\n\n"
+        "\u2022  Digite o nome de uma carta ou palavra-chave na caixa de busca.\n"
+        "\u2022  Clique em um resultado para visualiz\u00e1-lo no Inspetor de Cartas.\n"
+        "\u2022  Use \u201cAdicionar ao Principal\u201d ou \u201cAdicionar ao Sideboard\u201d para adicion\u00e1-la ao deck.\n"
+        "\u2022  Abra o Teclado de Mana para inserir rapidamente s\u00edmbolos de custo de mana.\n"
+        "\u2022  Use \u201cCopiar\u201d para copiar a lista de deck para a \u00e1rea de transfer\u00eancia."
+    ),
+    "tutorial.step5.title": "Guia de Sideboard",
+    "tutorial.step5.body": (
+        "A aba Guia de Sideboard permite registrar notas por matchup.\n\n"
+        "\u2022  Adicione uma entrada para cada arqu\u00e9tipo que voc\u00ea enfrenta.\n"
+        "\u2022  Registre as cartas para ENTRAR e SAIR em cada matchup.\n"
+        "\u2022  Marque as flex slots \u2014 cartas cuja quantidade varia por matchup.\n"
+        "\u2022  Fixe o guia para mant\u00ea-lo vis\u00edvel enquanto revisa outras abas.\n"
+        "\u2022  Exporte ou importe como CSV para compartilhar guias com colegas de equipe."
+    ),
+    "tutorial.step6.title": "Voc\u00ea Est\u00e1 Pronto!",
+    "tutorial.step6.body": (
+        "Esse foi o tour r\u00e1pido do MTGO Tools.\n\n"
+        "Mais algumas dicas:\n"
+        "\u2022  Use o menu \u2699 Configura\u00e7\u00f5es para carregar sua cole\u00e7\u00e3o MTGO, baixar imagens de cartas, "
+        "atualizar o banco de dados de cartas ou mudar o idioma.\n"
+        "\u2022  Notas de Deck permitem manter anota\u00e7\u00f5es em texto livre vinculadas a qualquer deck.\n"
+        "\u2022  O estado da sess\u00e3o (deck atual, formato, tamanho da janela) \u00e9 salvo automaticamente.\n\n"
+        "Boa sorte nas suas partidas!"
+    ),
+}
